Add MODULE.bazel to rules_jsonnet (#179)
* Make fewer assumptions about Bazel's pathname structure

When using bzlmod, there is no guarantee that the underlying name of the
external is "examples". This may currently cause these tests to fail due
to non-matching regular expressions.

* Give rules_jsonnet a MODULE.bazel

* Document how rules_jsonnet may be used with bzlmod

* Stop referencing WORKSPACE file in tests

* Let the examples repository use MODULE.bazel

Instead of using local_repository() inside of the root WORKSPACE, give
examples/ its own MODULE.bazel that uses local_repository() to access
the copy of rules_jsonnet at the repository's root.

This allows us to move the bazelrc and workspace_status command we have
at the root directory into the examples/ directory as well.

* Update CI to match reality

It looks like it's not easy to make the WORKSPACE based build work
properly, as jsonnet_go depends on copies of Gazelle and rules_go that
are far too old. At least make sure that the MODULE.bazel based builds
are tested.