Perform proper string concatenation in jsonnet_to_json() (#183)
It looks like #182 introduced a small regression, where '%' was used
instead of '+' to perform string concatenation. This went by unnoticed,
because we seemingly don't have any testing coverage for plain
jsonnet_to_json().

Solve this by duplicating one of the import tests that was affected by
this to invoke plain jsonnet_to_json().