Prepare for releasing rules_jsonnet 0.6.0 (#187)
* Automate the release process

This change imports a GitHub workflow that is similar to the one we have
in rules_python that gets launched when new tags are created. We can use
that to automatically perform releases on GitHub and attach a stable
source tarball.

Issue: #170

* Use zero version numbers in helper modules

These will always use the very latest version of the source code that's
present in this repository. There is no need to fill in an exact number
here.

* Bump the version number of the rules_jsonnet module to 0.6.0

I think that it makes sense to cut a new release of rules_jsonnet. First
and foremost because this will be the first official release to support
bzlmod. It also addresses the import path issue that prevents people
from using these rules in multi-workspace environments.